Move registry to main package
